Add support for multiple runs per dataset
A 'run' consists of a set of options which are passed to the python
script, which can be used to customize the set of plots to produce,
or change the cuts, or basically anything the user wants.

The plots for all the various runs are done within the same event
loop, so if only the cut string is changed, runtime penality should
be really small. Each set of plots for each run is saved in its own
file.

Run options are accessible from the global dict 'options' in the
script (read only). In addition, the whole current dataset is also
accessible from the global dict 'dataset'. This variable is writable,
which mean the python script can and should change the output name
of the current name with a custom naming scheme to avoid file
overwrite.