Add helper functions to combine scale-factors
A function `common::combineScaleFactors<>()` is available globally
for both `histFactory` and `treeFactory`.

Two signatures are available:
  - The first one takes an array of scale-factors, and the wanted
  variation, and return the combination of the scale-factors.
  - The second one is similar, but accepts also a correlation matrix
  between the different scale-factors.

The template is just the number of scale-factors to consider. The
algorithm is generic enough to deal with an arbitrary number.

For signature #1, non-correlated scale-factors are considered.

This function is only valid if the combination of scale-factors is a
simple product of the terms.