Add new 'normalize-to' option for plots
This inform histFactory on how to normalize the histogram. Default value
is 'nominal' and the histogram will be normalized to the sum of event
weight.

Valid values are anything contained in the input dataset JSON file, in the
'extras-event-weight-sum' dictionary.

Plots are now only rescaled if the input dataset is MC.