[HistFactory] add support for sample weights
The weights expressions must be defined in the python configuration,
through the `sample_weights` dictionnary. The key is an identifier, and
the value the expression of the weight.

In the sample json, one can use one of the key defined in the python
configuration for the new `sample-weight` option. The weight evaluated
from the expression will be used to scale all the histograms. If set to
`none` or not specified, a default weight of 1 is applied.